+++ ChangeLog.md
@@ -2,7 +2,7 @@
 
 ## Unreleased changes
 
-- Add endpoint `GET /v3/accTransactions/{account address}` get the transactions affecting an account's ccd or plt balances (including special transaction outcomes for suspended/inactive validators and including CCD transactions with memos).
+- Add endpoint `GET /v3/accTransactions/{account address}` that gets the transactions affecting an account's ccd or plt balances (including special transaction outcomes for suspended/inactive validators and including CCD transactions with memos).
 - Add endpoint `GET /v2/accBalance/{account address}` that includes the plt (protocol level token) balances of an account.
 - Add endpoint `GET /v0/plt/tokens` to get a list of all plt token infos.
 - Add endpoint `GET /v0/plt/tokenInfo/{tokenId}` to get the token info about a given plt token and its module state (with metadata url)